hello api

+++ backend/src/main/java/com/osathi_java_example/demo/repository/HelloRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.osathi_java_example.demo.model.Hello;
 import java.util.List;
 
-public interface HelloRepository extends JpaRepository<Hello, Long> {
+public interface HelloRepository extends JpaRepository<Hello, Integer> {
     List<Hello> findAll();
 }
